Inline the rest of box_region
